try to fix image issue

+++ src/components/blog/Blog.js
@@ -33,7 +33,9 @@ function Blog({
               buttonValue="Delete"
               iconClass="fas fa-trash"
               handleClick={() => {
-                handleDelete(postId);
+                if (confirm("Do you want to Delete this Blog?")) {
+                  handleDelete(postId);
+                }
               }}
             />
           </div>
